+++ whatsonpypi/client.py
@@ -37,7 +37,7 @@ class WoppResponse(object):
 
     @property
     def homepage(self):
-        return self.json.get('home_page')
+        return self.json.get('homepage')
 
     @property
     def package_url(self):
@@ -46,6 +46,10 @@ class WoppResponse(object):
     @property
     def project_urls(self):
         return self.json.get('project_urls')
+
+    @property
+    def project_docs(self):
+        return self.project_urls.get('Documentation') or self.homepage
 
     @property
     def requires_python(self):
@@ -103,16 +107,17 @@ class WoppClient(object):
         # default_hooks() returns {'response': []}
         self.request_hooks = request_hooks or hooks.default_hooks()
 
-    def request(self, package=None, timeout=3.1, max_retries=3, ):
+    def request(self, package=None, version=None, timeout=3.1, max_retries=3):
         """
         Make a HTTP GET request with the provided params
 
         :param timeout: request timeout seconds
         :param max_retries: number of times to retry on failure
         :param package: name of the python package to search
+        :param version: version of the python package to search
         :return: response serialized by WoppResponse object
         """
-        url = self._build_url(package)
+        url = self._build_url(package, version)
         req_kwargs = {
             'method': 'GET',
             'url': url,
@@ -143,20 +148,26 @@ class WoppClient(object):
         )
 
         if response.status_code == 404:
-            raise PackageNotFoundError("Sorry, but that package couldn't be found on PyPI.")
+            raise PackageNotFoundError("Sorry, but that package/version couldn't be found on PyPI.")
 
         # serialize response
         wopp_response = WoppResponse(int(response.status_code), response.cleaned_json)
         return wopp_response
 
-    def _build_url(self, package=None):
+    def _build_url(self, package=None, version=None):
         """
         Builds the URL with the path params provided.
 
         :param package: name of package
+        :param version: version of package
         :return: fully qualified URL
         """
         if package is None:
             raise PackageNotProvidedError('A package name is needed to proceed.')
 
-        return "{}/{}/json".format(self.base_url, package)
+        if version is not None:
+            url = "{}/{}/{}/json".format(self.base_url, package, version)
+        else:
+            url = "{}/{}/json".format(self.base_url, package)
+
+        return url