+++ whatsonpypi/exceptions.py
@@ -24,3 +24,21 @@ class PackageNotFoundError(WoppException):
     """
     Raised when a package is not found on PyPI
     """
+
+
+class DocsNotFoundError(WoppException):
+    """
+    Raised when a package does not have documentation or homepage urls
+    """
+
+
+class URLLaunchError(WoppException):
+    """
+    Raised when there's a problem opening a URL in the browser
+    """
+
+
+class BadRequirementsFormatError(WoppException):
+    """
+    Raised when the requirements file is not in the right format.
+    """